refactor: separate business logic from CLI
